chore: move clickaway directive to its own file
