+++ solution/ui/regulations/eregs-component-lib/src/components/tooltips/TooltipContainer.vue
@@ -82,41 +82,6 @@ const leftWarning = (el) => el.getBoundingClientRect().left < 130;
 export default {
     name: "TooltipContainer",
 
-    components: {},
-
-    // https://www.vuesnippets.com/posts/click-away/
-    // https://dev.to/jamus/clicking-outside-the-box-making-your-vue-app-aware-of-events-outside-its-world-53nh
-    directives: {
-        clickaway: {
-            bind(el, { value }) {
-                if (typeof value !== "function") {
-                    console.warn(`Expect a function, got ${value}`);
-                    return;
-                }
-
-                const clickawayHandler = (e) => {
-                    const elementsOfInterest = Array.from(
-                        el.parentElement.children
-                    );
-                    const clickedInside = elementsOfInterest.filter((el) =>
-                        el.contains(e.target)
-                    );
-                    return clickedInside.length || value();
-                };
-
-                el.__clickawayEventHandler__ = clickawayHandler;
-
-                document.addEventListener("click", clickawayHandler);
-            },
-            unbind(el) {
-                document.removeEventListener(
-                    "click",
-                    el.__clickawayEventHandler__
-                );
-            },
-        },
-    },
-
     props: {
         btnClass: {
             type: String,
